Request body: discarded body now treated as no body.
Notably, proxying of such requests now uses no Content-Length instead
of "Content-Length: 0", and the $content_length variable is empty (instead
of "0").

This might be beneficial from correctness point of view, since requests
with discarded body, such as during processing of error pages, do not pretend
there is a zero-length body, but instead do not contain body at all.  For
example, this might be important for PUT requests, where a zero-length
body could be incorrectly interpreted as a real request body.

This also slightly simplifies the code.