Core: PID file writing synchronization.
Now, ngx_daemon() does not call exit() in the parent process immediately,
but instead waits for the child process to signal it actually started (and
wrote the PID file if configured to).  This ensures that the PID file
already exists when the parent process exits.

To make sure that signal handlers won't cause unexpected logging in the
parent process if the child process dies (for example, due to errors when
writing the PID file), ngx_init_signals() is moved to the child process.

This resolves "PID file ... not readable (yet?) after start" and "Failed
to parse PID from file..." errors as observed with systemd.

Note that the errors observed are considered to be a bug in systemd, which
isn't able to work properly with traditional Unix daemons.  Still, the
workaround is implemented to make sure there will be no OS vendor patches
trying to address this.