Core: better command-line help.
New users have no idea what kind of command-line arguments the program uses.
Asking for help with a long '--help' option is a common strategy, and it's easy
to implement.

Also, passing some nonsense after the help request will cause "angie" to report
an error instead of providing help.  With this change, the help request is
given the highest priority and is returned immediately after it is recognized.

Passing an incomplete option ("-") is currently ignored, while it is an error.
Appearance of such a symbol may indicate a typical error, when some shell
variable with option value was not set with only the separator being left.
Reporting an error in that case may be better than silently ignoring strange
input.