Avoided updating of parent PID value after daemonization.
The "ngx_parent" value is only used to detect conditions where the daemon is in
a binary upgrade procedure and the old master process hasn't exited yet.

To do this, it's checked for equality with ngx_getppid().  As a result, if it's
set to the parent process during daemonization, the binary upgrade signal is
ignored with an error: "[crit] the changing binary signal is ignored: you
should shutdown or terminate before either old or new binary's process" until
the parent process is terminated, which may not happen immediately after forking
under load on a slow system.

On the other hand, it should be absolutely safe to upgrade the binary before
the process left after daemonization has finally exited.

Notably, this problem has been observed sporadically in the "binary_upgrade.t"
test, as it sends a USR2 signal immediately after starting the daemon and
checking its pid file.