Core: changed ngx_buf_tag_t to uintptr_t.
There is no real difference, but "void *" as previously used results
in "ISO C forbids conversion of function pointer to object pointer type"
warnings with -Wpedantic when a function pointer is used as a tag.

Changing the type to uintptr_t makes the conversion always valid, since
any pointer type can be converted to an integer type.