conf/mime.types: add image/jxl internet media type
image/jxl for JPEG XL images is an officially reocgnized
internet media type[1] by the IANA, and browser support
is starting to exist (e.g. safari on macOS and iOS, plus
Firefox behind a nightly flag, etc.) so servers should be
able to serve .jxl images out of the box.

[1]: https://www.iana.org/assignments/media-types/image/jxl
[2]: https://web.archive.org/web/20240306204210/
       https://www.iana.org/assignments/media-types/image/jxl