Fixed binary upgrade with systemd after fb712e5081ff.
Using PR_SET_CHILD_SUBREAPER makes systemd a subreaper for a running process.
As a result, the getppid() call after daemonization returns the PID of the
process that originally started the master process, so after fb712e5081ff it
appears to be equal to "ngx_parent", which causes the USR2 signal to be ignored.

The fix is to explicitly set "ngx_parent" to an invalid PID value, which cannot
be equal to any possible result of the getppid() call.