Stream: TLS early data support.
The "ssl_early_data" directive controls if early data is enabled.

The feature is disabled by default, as it may have security implications.
This should only be enabled with applications that are aware.

Note that the feature has a limited use in stream: there is no means currently
to get and pass early data status to backend.